Moved MessageRateLimiter to its own file.
Added lib folder where all of our files can go.
